+++ src/Controller/FrontController.php
@@ -24,7 +24,7 @@ class FrontController {
 	{
 		$path = rtrim($_SERVER['REQUEST_URI'], '/');
 
-		if( !empty($path) && ($path == WP_FOLDER || $path == '/web'.WP_FOLDER) ){
+		if( !empty($path) && ($path == WP_FOLDER || $path == PUBLIC_DIR.WP_FOLDER) ){
 
 			wp_redirect(is_user_logged_in() ? admin_url() : wp_login_url());
 
@@ -107,11 +107,7 @@ class FrontController {
 	private function loadConfig()
 	{
 		global $_config;
-
 		$this->config = $_config;
-
-		self::$domain_name      = $this->config->get('domain_name', 'app');
-		self::$languages_folder = WP_CONTENT_DIR . '/languages';
 	}