+++ src/Plugin/EditorPlugin.php
@@ -48,9 +48,11 @@ class EditorPlugin {
 
 		$wp_admin_bar->remove_node('themes');
 		$wp_admin_bar->remove_node('updates');
-		$wp_admin_bar->remove_node('comments');
-		$wp_admin_bar->remove_node('wp-logo');
-	}
+        $wp_admin_bar->remove_node('wp-logo');
+
+        if( in_array('edit-comments.php', $this->config->get('remove_menu_page', [])) )
+            $wp_admin_bar->remove_node('comments');
+    }
 
 	/**
 	 * Filter admin menu entries